+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hard levels are now either fully hidden, or hidden in checkerboard going from level 102 and up
 - Special levels can now contain checkerboard hidden blocks from level 200 going up
 
+### Changed
+
+- Rebalanced AutoMove
+
 ## Version 0.6.1 - 2024-09-27
 
 ### Added